Fix ADC sampling
- Fixed ADC_Read function
- The path to the ADC data files is generally different (hooray)
- Modified run.sh to work it out and pass to the program
  - run.sh also loads kernel modules, etc

+++ server/bbb_pin.h
@@ -24,7 +24,7 @@ extern void ADC_Unexport();
 extern bool GPIO_Read(int pin);
 extern void GPIO_Set(int pin, bool value);
 
-extern long ADC_Read(int pin);
+extern int ADC_Read(int pin);
 
 extern void PWM_Set(int pin, bool polarity, long period, long duty); // period and duty are in ns
 extern void PWM_Stop(int pin);
